requirements.txt
amqp==5.2.0
asgiref==3.4.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==3.2.25
djangorestframework==3.15.1
gunicorn==21.2.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
kombu==5.1.0
packaging==21.3
prompt-toolkit==3.0.36
psycopg2-binary==2.9.8
pyparsing==3.1.2
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.1.1
vine==5.1.0
wcwidth==0.2.13
zipp==3.6.0
